pyproject.toml
[project]
name = "zepr-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "rajeshrc12",email = "rajesh.charhajari@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=7.0.0,<8.0.0)",
    "authlib (>=1.6.5,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "starlette (>=0.48.0,<0.49.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "app.run:main"

[[tool.poetry.packages]]
include = "app"
